requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
asyncpg==0.27.0
python-dotenv==1.0.0
greenlet==2.0.2
pydantic==1.10.2
email-validator==1.3.0
psycopg2==2.9.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
python-multipart==0.0.9
